requirements.txt
Flask
google-generativeai
replicate
requests
Pillow
python-dotenv
